+++ cmd/hidden_lake/composite/service_traffic_messenger/cmd/hlc_stm/init_app.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"context"
-	"errors"
-	"sync"
+	"fmt"
 
 	hlm_app "github.com/number571/go-peer/cmd/hidden_lake/messenger/pkg/app"
 	hls_app "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/app"
 	hlt_app "github.com/number571/go-peer/cmd/hidden_lake/traffic/pkg/app"
+	"github.com/number571/go-peer/pkg/state"
 	"github.com/number571/go-peer/pkg/types"
 	"github.com/number571/go-peer/pkg/utils"
 )
@@ -17,8 +17,7 @@ var (
 )
 
 type sApp struct {
-	fMutex sync.Mutex
-	fIsRun bool
+	fState state.IState
 	fHLS   types.IRunner
 	fHLT   types.IRunner
 	fHLM   types.IRunner
@@ -41,31 +40,21 @@ func initApp(pPath, pKey string) (types.IRunner, error) {
 	}
 
 	return &sApp{
-		fHLS: hlsApp,
-		fHLT: hltApp,
-		fHLM: hlmApp,
+		fState: state.NewBoolState(),
+		fHLS:   hlsApp,
+		fHLT:   hltApp,
+		fHLM:   hlmApp,
 	}, nil
 }
 
 func (p *sApp) Run(pCtx context.Context) error {
-	err := func() error {
-		p.fMutex.Lock()
-		defer p.fMutex.Unlock()
-
-		if p.fIsRun {
-			return errors.New("application already running")
-		}
-
-		p.fIsRun = true
-		return nil
-	}()
-	if err != nil {
-		return err
+	if err := p.fState.Enable(nil); err != nil {
+		return fmt.Errorf("application running error: %w", err)
 	}
 	defer func() {
-		p.fMutex.Lock()
-		p.fIsRun = false
-		p.fMutex.Unlock()
+		if err := p.fState.Disable(nil); err != nil {
+			panic(err)
+		}
 	}()
 
 	var (