+++ cmd/hidden_lake/messenger/pkg/app/app.go
@@ -10,8 +10,8 @@ import (
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/config"
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/database"
 	"github.com/number571/go-peer/pkg/logger"
+	"github.com/number571/go-peer/pkg/state"
 	"github.com/number571/go-peer/pkg/types"
-	"github.com/number571/go-peer/pkg/utils"
 
 	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/messenger/pkg/settings"
 	"github.com/number571/go-peer/internal/interrupt"
@@ -24,7 +24,7 @@ var (
 )
 
 type sApp struct {
-	fIsRun bool
+	fState state.IState
 	fMutex sync.Mutex
 
 	fConfig config.IConfig
@@ -47,6 +47,7 @@ func NewApp(
 	stdfLogger := std_logger.NewStdLogger(pCfg.GetLogging(), std_logger.GetLogFunc())
 
 	return &sApp{
+		fState:      state.NewBoolState(),
 		fConfig:     pCfg,
 		fPathTo:     pPathTo,
 		fHTTPLogger: httpLogger,
@@ -55,26 +56,24 @@ func NewApp(
 }
 
 func (p *sApp) Run(pCtx context.Context) error {
-	err := func() error {
-		p.fMutex.Lock()
-		defer p.fMutex.Unlock()
-
-		if p.fIsRun {
-			return errors.New("application already running")
-		}
-
+	enableFunc := func() error {
 		if err := p.initDatabase(); err != nil {
 			return fmt.Errorf("init database: %w", err)
 		}
-
 		p.fStdfLogger.PushInfo(fmt.Sprintf("%s is running...", pkg_settings.CServiceName))
-		p.fIsRun = true
-
 		return nil
-	}()
-	if err != nil {
-		return err
 	}
+	if err := p.fState.Enable(enableFunc); err != nil {
+		return fmt.Errorf("application running error: %w", err)
+	}
+
+	defer func() {
+		disableFunc := func() error {
+			p.fStdfLogger.PushInfo(fmt.Sprintf("%s is shutting down...", pkg_settings.CServiceName))
+			return p.stop()
+		}
+		_ = p.fState.Disable(disableFunc)
+	}()
 
 	p.initIncomingServiceHTTP()
 	p.initInterfaceServiceHTTP()
@@ -112,21 +111,15 @@ func (p *sApp) Run(pCtx context.Context) error {
 
 	select {
 	case <-pCtx.Done():
-		return p.stop()
+		return nil
 	case err := <-chErr:
-		return utils.MergeErrors(
-			fmt.Errorf("got run error: %w", err),
-			p.stop(),
-		)
+		return fmt.Errorf("got run error: %w", err)
 	}
 }
 
 func (p *sApp) stop() error {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
-
-	p.fStdfLogger.PushInfo(fmt.Sprintf("%s is shutting down...", pkg_settings.CServiceName))
-	p.fIsRun = false
 
 	err := interrupt.CloseAll([]types.ICloser{
 		p.fIntServiceHTTP,