+++ pkg/state/state_test.go
@@ -1,0 +1,5 @@
+package state
+
+import "testing"
+
+func TestNothing(t *testing.T) {}