+++ pkg/state/types.go
@@ -1,0 +1,8 @@
+package state
+
+type IStateFunc func() error
+
+type IState interface {
+	Enable(IStateFunc) error
+	Disable(IStateFunc) error
+}